Moave ProducerRecords constructors to package object
